+++ internal/acceptance_tests/blocks/acl_entries_many.tf
@@ -1,5 +1,5 @@
 resource "fastly_service_cdn_acl_entries" "test" {
   service_id     = fastly_service_cdn.test.id
-  acl_id         = fastly_service_acl.test.acl_id
+  acl_id         = fastly_service_cdn_acl.test.acl_id
   manage_entries = true
 {{.ENTRIES}}}
