+++ internal/acceptance_tests/blocks/acl_entries_two.tf
@@ -1,6 +1,6 @@
 resource "fastly_service_cdn_acl_entries" "test" {
   service_id     = fastly_service_cdn.test.id
-  acl_id         = fastly_service_acl.test.acl_id
+  acl_id         = fastly_service_cdn_acl.test.acl_id
   manage_entries = true
 
   entry {
