+++ templates/resources/service_cdn_acl_entries.md.tmpl
@@ -28,7 +28,7 @@ resource "fastly_service_cdn" "example" {
   }
 }
 
-resource "fastly_service_acl" "example" {
+resource "fastly_service_cdn_acl" "example" {
   service_id = fastly_service_cdn.example.id
   version    = 1
   name       = "example_acl"
@@ -36,7 +36,7 @@ resource "fastly_service_acl" "example" {
 
 resource "fastly_service_cdn_acl_entries" "example" {
   service_id     = fastly_service_cdn.example.id
-  acl_id         = fastly_service_acl.example.acl_id
+  acl_id         = fastly_service_cdn_acl.example.acl_id
   manage_entries = true
 
   entry {
